Fix warning

+++ driver/buff_alloc.c
@@ -61,8 +61,8 @@ static void print_genpool_cb(struct gen_pool *pool, struct gen_pool_chunk *chunk
 
 void knacs_buff_alloc_print(struct gen_pool *pool)
 {
-    pr_info("Total Size: %ld, Total Avail: %ld\n", gen_pool_size(pool),
-            gen_pool_avail(pool));
+    pr_info("Total Size: %ld, Total Avail: %ld\n", (long)gen_pool_size(pool),
+            (long)gen_pool_avail(pool));
     gen_pool_for_each_chunk(pool, print_genpool_cb, NULL);
 }
